requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
deep-translator
python-dotenv